changing filenames to reflect init vs. maintain status of correaltion
